pyproject.toml
[project]
name = "ai-horde-oai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
tui = ["textual>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "pytest-cov>=6.0",
    "httpx>=0.28",
]

[project.scripts]
horde-oai = "app.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
